modify max_expansion variable type to long

+++ src/fs/core/search/drivers/breadth_first_search.cxx
@@ -30,7 +30,7 @@ BreadthFirstSearchDriver<StateModelT>::search(Problem& problem, const Config& co
 	auto model = setup(problem);
 //	int maxw = config.getOption<int>("width.max", -1);
 	auto stop_on_goal = config.getOption<bool>("until_first_goal", false);
-	auto max_exp = config.getOption<int>("max_expansions", -1);
+	auto max_exp = config.getOption<long>("max_expansions", -1);
 
 	OpenListT queue = OpenListT();
 	EventUtils::setup_stats_observer<NodeT>(_stats, _handlers);
